fix buffer allocations for readStream
The length in real valued elements seems to be an rtl hold over,
but soapysdr uses a length in number of samples for its calls.
In many cases the buffers were smaller than the size being passed to read stream.
This change makes use of SoapySDR_formatToSize() to replace sizeof(types).